Refactor vizStore and selectedGeographyStore
