+++ src/components/Layout.svelte
@@ -3,6 +3,14 @@
   import Footer from "./Footer.svelte";
   import Map from "./Map.svelte";
   import MapKey from "./MapKey.svelte";
+  import { page } from "$app/stores";
+  import { getSelectedGeography } from "../helpers/categoryHelpers";
+  import { setSelectedGeographyStore } from "../data/setSelectedGeographyStore";
+
+  $: if ($page) {
+    const g = getSelectedGeography($page.url);
+    setSelectedGeographyStore({ geoCode: g.geoCode, geoType: g.geoType });
+  }
 </script>
 
 <div class="xl:tw-absolute tw-inset-0 xl:tw-flex tw-flex-col">
