+++ src/components/LocationPage.svelte
@@ -3,32 +3,10 @@
   import Heading from "./Heading.svelte";
   import LocationOverview from "./LocationOverview.svelte";
   import { selectedGeographyStore } from "../stores/stores";
-  import { fetchSelectedGeographyData, parseSelectedGeographyData } from "../data/fetchVizData";
-  import { getCodesForCategory } from "../helpers/categoryHelpers";
-
-  let householdsCodes = getCodesForCategory("housing", "size-of-household", "default", "1-person-households");
-  let peopleCodes = getCodesForCategory("health", "general-health", "default", "very-good-health");
-  let allHouseholdsTotal = "";
-  let allPeopleTotal = "";
 
   $: selectedGeographyDisplayName = $selectedGeographyStore?.displayName;
-
-  $: if ($selectedGeographyStore.geoCode) {
-    fetchSelectedGeographyData({
-      totalCode: householdsCodes.totalCode,
-      categoryCodes: householdsCodes.categoryCodes,
-      geoCode: $selectedGeographyStore.geoCode,
-    }).then((table) => {
-      allHouseholdsTotal = parseSelectedGeographyData(table, householdsCodes.totalCode).selectedGeographyTotal;
-    });
-    fetchSelectedGeographyData({
-      totalCode: peopleCodes.totalCode,
-      categoryCodes: peopleCodes.categoryCodes,
-      geoCode: $selectedGeographyStore.geoCode,
-    }).then((table) => {
-      allPeopleTotal = parseSelectedGeographyData(table, peopleCodes.totalCode).selectedGeographyTotal;
-    });
-  }
+  $: allHouseholdsTotal = $selectedGeographyStore?.allHouseholdsTotal;
+  $: allPeopleTotal = $selectedGeographyStore?.allPeopleTotal;
 </script>
 
 <Heading serviceTitle={selectedGeographyDisplayName ? selectedGeographyDisplayName : "England and Wales"} />
