--- src/data/fetchGeographyData.ts
@@ -1,15 +1,0 @@
-import { englandAndWales } from "../helpers/spatialHelper";
-
-const apiBaseUrl = `https://cep5lmkia0.execute-api.eu-west-1.amazonaws.com/dev`;
-
-export const fetchGeographyLookup = async (location: string, useCode = true) => {
-  // return englandAndWales if the name or code matches...
-  if ((useCode && englandAndWales.meta.code == location) || englandAndWales.meta.name == location) {
-    return JSON.stringify(englandAndWales);
-  }
-  // otherwise query API
-  const url = `${apiBaseUrl}/geo/2011?${useCode ? "geocode" : "geoname"}=${location}`;
-  const response = await fetch(url);
-  const data = await response.text();
-  return data;
-};
