+++ src/helpers/spatialHelper.ts
@@ -1,10 +1,10 @@
-import type { Bbox, GeographyLookupProps } from "../types";
+import type { Bbox, GeographyInfo } from "../types";
 
 export const getBboxString = (args: Bbox) => {
   return [args.east, args.north, args.west, args.south].map((n) => n.toFixed(6)).join(",");
 };
 
-export const englandAndWales: GeographyLookupProps = {
+export const englandAndWales: GeographyInfo = {
   meta: {
     name: "England and Wales",
     code: "K04000001",
@@ -12,6 +12,12 @@ export const englandAndWales: GeographyLookupProps = {
   },
   geo_json: {
     type: "FeatureCollection",
-    features: [],
+    features: [
+      {
+        id: "bbox",
+        type: "blah",
+        geometry: { type: "blah", coordinates: [2.08, 55.68, -6.59, 48.53] },
+      },
+    ],
   },
 };
