+++ src/routes/__layout.svelte
@@ -1,16 +1,8 @@
 <script>
-  import { page } from "$app/stores";
-  import { selectedGeographyStore } from "../stores/stores";
   import "mapbox-gl/dist/mapbox-gl.css";
-  import "../i18n/i18n.ts";
-  import updateSelectedGeography from "../helpers/updateSelectedGeography";
-  import Layout from "../components/Layout.svelte";
-
   import "../app.css";
-
-  $: if ($page) {
-    updateSelectedGeography($page.url.search, $selectedGeographyStore);
-  }
+  import "../i18n/i18n.ts";
+  import Layout from "../components/Layout.svelte";
 </script>
 
 <Layout>
