+++ src/types.ts
@@ -26,9 +26,11 @@ export type VizData = {
 
 export type SelectedGeographyData = {
   geoType: GeoType;
-  displayName: string;
   geoCode: string;
-  bbox?: [number, number, number, number];
+  displayName: string;
+  bbox: [number, number, number, number];
+  allHouseholdsTotal: number;
+  allPeopleTotal: number;
 };
 
 export type GeographyAutoSuggestProps = {
@@ -40,16 +42,16 @@ export type GeographyAutoSuggestProps = {
   sanitisedText?: string;
 };
 
-export type GeographyLookupProps = {
+export type GeographyInfo = {
   meta: {
     name: string;
     code: string;
     geotype: string;
   };
-  geo_json: GeographyGeo;
+  geo_json: GeographyGeojson;
 };
 
-type GeographyGeo = {
+type GeographyGeojson = {
   type: string;
   features: GeographyFeatures[];
 };
